Added CallThreshold to config

+++ APIUsage.Core/CalculateCostResponse.cs
@@ -6,8 +6,8 @@ namespace APIUsage.Core
 {
     public class CalculateCostResponse
     {
-        public virtual decimal MonthlyCharge { get; set; }
-        public virtual int TotalNoOfCalls { get; set; }
+        public virtual double MonthlyCharge { get; set; }
+        public virtual double TotalNoOfCalls { get; set; }
         public virtual string ResponseMessage { get; set; }
         public virtual bool IsSuccessful { get; set; }
     }
